MWE iceflow updated
Small update using SciMLSensitivity instead of DiffEqFlux
